replacing old ObjText with the new one
